+++ larq_compute_engine/tflite/tests/bconv2d_op_model.h
